finished batch 3 and fixed naming schemes for prev programs

+++ batch_2_python_programs/print_even_number_count.py
@@ -1,7 +1,7 @@
 def input_num (prompt):
     while True:
         try: 
-            num = int(input(prompt))
+            num = float(input(prompt))
             return num
         except ValueError:
             print("enter numbers only")
